requirements.txt
pytest==7.4.3
pytest-flask==1.2.0
pytest-mock==3.12.0
requests==2.31.0